feat: everything
A lot of stuff to clean up/fix and all of the docs to rewrite, but
I have been working for almost a month on this refactor, so it's
probably a good idea to actually commit my work so far.

For my minimal testing, it seems like most components are working
as intended, although there's a few kinks to work out here and there,
like help text formatting or params arguments.