chore: alternative to Environment.Version in DocParser
We basically do a reflection check to see if StringSplitOptions.TrimEntries
exists, and call the appropriate function.

According to some quick benchmarks I did, the speedup from using
the right options outweighs the reflection cost after under 10 calls
to TrimAndSplit, and even then the length of the input string and
the number of newlines means that even for just one call, we turn a
profit around 500 characters; the equilibrium is probably somewhere
between (l=10, n=10) and (l=500, n=1).