feat!: don't check for implicit castability to string in ParserFinder
Types that are implictely castable to `string` WON'T be seen as a
case of ParserInfo.Identity. This was a difficult decision, but in
all honesty I don't think this is *that* common of a pattern, so I
think it's a worthy sacrifice. If anybody needs it, they can just
open an issue.

Fixes #14