feat: refactor CmdDesc and gen'd code
This started out as a fix for #18, but I ended up having to change
quite a lot of things in the generated code to get it to work
correclty. I'm not suuuuper happy with everything that changed here,
like the fact that a lot of code that was previously in the static
ReclineProgram.g.cs file ended up repeated in descriptors' methods,
or that we had to take `args` and `argsIdx` out into the open. I
might try to fix some of that later, but for now this works and
it's pretty cool B)

(and it's one more step towards a single enormous be-all-end-all
Main function :P)

Fixes #18