fix: fix xml comments not affecting help text
For Roslyn to actually read the xml docs and return them when calling
`GetDocumentationCommentXml`, you need to set the project property
`GenerateDocumentationFile`, but that enables warnings CS1587, CS1591
and CS1573 by default. Since we don't want to suddenly submerge our
users in a deluge of warnings *but* also don't want to disable any
warning the user actually wanted, in case they already set the
aforementioned property.

The solution is to do all the property magic in the .targets file:
that enables us to check whether the user enabled it manually or not.
This way, we only disable the warnings in case the user *didn't* set
`GenDocFile` to true. Then, we force the value of that prop to true,
and we're done!