fix: fix interference between default cmds and params
In the case of a default command having multiple positional/params args,
groups' TryUpdateCommand would throw an error, but only for the 2nd arg.

This is because there's a check making sure that no argument was given
before transitioning to a new command/group. However, it seems like the
only scenario where we might have a non-empty arg count is when using
default commands.

The only precaution we need to take is to avoid invoking a sub-command
if we've already processed some args, as that locks us into the default
command anyway: if those weren't the default command's args, then
they're just orphaned, since there's no actual command to handle them,
as we're inside a group right now, and a group only ever allows args
to be added in case it has a default command. Otherwise, ArgCount
should always be zero while we're still inside a non-invokable group.

The solution here is to silently reject any call to TryUpdateCommand
when ArgCount != 0. No need for a runtime error, since it's mostly
just an ambiguity brought about by the design and use of default cmds,
which as we just saw can be resolved pretty easily once we realise
other interpretations don't really make sense.