feat: allow parsers/validators to be generic methods (with arity = 1)
This is a part of the first solution from the thread in #43. It
partially solves the problem of generic methods being straight up
unusable as parsers or validators, by always assuming that their
generic parameter is the type of the target symbol.

This is the most basic version, which just dumbly substitutes directly
the type, without ever checking if it is compatible. It also doesn't
special-case repeatable options for validators, so it isn't aware
that it should try item-wise substitution in the case of arrays.